Move lang utils to util
