+++ taichi/codegen/cc/cc_runtime.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "taichi/lang_util.h"
+#include "taichi/util/lang_util.h"
 
 TLANG_NAMESPACE_BEGIN
 class CCProgramImpl;
