+++ taichi/codegen/cc/cc_utils.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "taichi/lang_util.h"
+#include "taichi/util/lang_util.h"
 #include "taichi/common/core.h"
 #include <fstream>
 #include <sstream>
