+++ taichi/codegen/cc/codegen_cc.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "taichi/lang_util.h"
+#include "taichi/util/lang_util.h"
 #include "taichi/codegen/codegen.h"
 
 TLANG_NAMESPACE_BEGIN
