+++ taichi/codegen/cc/context.h
@@ -40,7 +40,7 @@ struct Ti_Context {
 #else  // _CC_INSIDE_KERNEL
 
 #include "cc_program.h"
-#include "taichi/lang_util.h"
+#include "taichi/util/lang_util.h"
 
 TLANG_NAMESPACE_BEGIN
