+++ taichi/jit/jit_session.h
@@ -4,7 +4,7 @@
 #include <functional>
 
 #include "taichi/runtime/llvm/llvm_fwd.h"
-#include "taichi/lang_util.h"
+#include "taichi/util/lang_util.h"
 #include "taichi/jit/jit_module.h"
 
 TLANG_NAMESPACE_BEGIN
