+++ taichi/program/callable.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "taichi/lang_util.h"
+#include "taichi/util/lang_util.h"
 
 namespace taichi {
 namespace lang {
