+++ taichi/program/kernel.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "taichi/lang_util.h"
+#include "taichi/util/lang_util.h"
 #include "taichi/ir/snode.h"
 #include "taichi/ir/ir.h"
 #include "taichi/rhi/arch.h"
