+++ taichi/util/file_sequence_writer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "taichi/lang_util.h"
+#include "taichi/util/lang_util.h"
 #include "taichi/ir/transforms.h"
 #ifdef TI_WITH_LLVM
 #include "taichi/runtime/llvm/llvm_fwd.h"
