+++ taichi/util/str.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <functional>
 
-#include "taichi/lang_util.h"
+#include "taichi/util/lang_util.h"
 
 TLANG_NAMESPACE_BEGIN
